perf(function):  Batch serialization with preserve encoding option support in remote functions (#15643)
Summary:


I noticed that the constant encodings were not preserved when we are receiving thrift request from Presto for remote function calls. We use a lot of constant parameters in our remote funciton udf and if we don't preserve the constant encoding, it will more than double the payload size.
After debugging the issue, the fix is to use batch serializer (instead of iterative serializer) which preserves the encoding correctly. I noticed 60% reduced in payload size after testing this change locally for our remote function udf. This is just a simple case, in other cases we will have even more savings (more than 3x reduction in payload size).

Reviewed By: pedroerp

Differential Revision: D87873415